refactor(location-manager): split Payload API contracts
