+++ apps/location-manager/packages/server/src/features/locations/services/integrations/clients/payload/payload-auth.types.ts
@@ -1,0 +1,19 @@
+export interface PayloadAuthResponse {
+  message: string;
+  user: {
+    id: string;
+    email: string;
+    role: string;
+  };
+  token: string;
+  exp: number;
+}
+
+export interface PayloadCustomAuthResponse {
+  token: string;
+  user?: {
+    id?: string;
+    email?: string;
+    role?: string;
+  };
+}
