+++ apps/location-manager/packages/server/src/features/locations/services/integrations/clients/payload/payload-instagram.types.ts
@@ -1,0 +1,23 @@
+export interface PayloadInstagramPostData {
+  title: string;
+  embedCode: string;
+  previewImage?: string;
+  status: "draft" | "published";
+}
+
+export interface PayloadInstagramPostResponse {
+  message: string;
+  doc: {
+    id: string;
+    title: string;
+    embedCode: string;
+    previewImage: {
+      id: string;
+      filename: string;
+      url: string;
+    };
+    status: "draft" | "published";
+    createdAt: string;
+    updatedAt: string;
+  };
+}
