+++ apps/location-manager/packages/server/src/features/locations/services/integrations/clients/payload/payload-locations.client.ts
@@ -5,7 +5,7 @@ import type {
   PayloadLocationCreateData,
   PayloadLocationCreateResponse,
   PayloadLocationQueryResponse,
-} from "./payload-api.types";
+} from "./payload-location.types";
 
 export class PayloadLocationsClient {
   constructor(private readonly authClient: PayloadAuthClient) {}
