+++ apps/location-manager/packages/server/src/features/locations/services/integrations/clients/payload/payload-media.types.ts
@@ -1,0 +1,16 @@
+export interface PayloadMediaAssetResponse {
+  message: string;
+  doc: {
+    id: string;
+    filename: string;
+    mimeType: string;
+    filesize: number;
+    width: number;
+    height: number;
+    url: string;
+    altText?: string;
+    location?: string;
+    createdAt: string;
+    updatedAt: string;
+  };
+}
