+++ apps/location-manager/packages/server/src/features/locations/services/integrations/clients/payload/payload-shared.types.ts
@@ -1,0 +1,9 @@
+export type PayloadCollection =
+  | "dining"
+  | "accommodations"
+  | "attractions"
+  | "nightlife"
+  | "key-locations"
+  | "tours";
+
+export type PayloadRelationshipId = string | number;
